Wedding-crm-main/app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    # Database settings
    DATABASE_TYPE = os.getenv("DATABASE_TYPE", "json")  # json or mongodb
    MONGO_URI = os.getenv("MONGO_URI")
    DATABASE_NAME = os.getenv("DATABASE_NAME")
    JSON_DATA_DIR = os.getenv("JSON_DATA_DIR", "data")
    
    # JWT settings
    JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY", "your-secret-key-change-in-production")
    JWT_ALGORITHM = "HS256"
    ACCESS_TOKEN_EXPIRE_MINUTES = 30

settings = Settings()